docs(api): Updated examples to use ES2015
Example code has been migrated from ES5 to ES2015. A consequence of moving towards ES2015 as a
default is that the default format for iterating over objects has changed. The documentation for how
to iterate objects also has examples, but these will not be updated until the default is changed.
Issue #4.