feat(transformation): transduce can now take a function as a reducer
In all of the theory, transduce takes a function for a reducer. But the actual function required a
reducer object and wouldn't accept a plain function. This forced use of toReducer to make a reducer
object, but there's no reason why transduce couldn't take care of that itself. Now it does. Issue
#16.

closes #16